ci: Re-organize "deploy-download-preview" build to include GitHub status check
Update the GitHub workflow to explicitly build a preview of the download
site and force push the built site onto the "deploy-download-preview"
branch.

The "commit-status-updater" Github action is used to add the status
check "netlify/deploy-download-preview" set first to "pending" with
the URL "netlify/deploy-download-preview" and then to "success" with
the URL "https://deploy-download-preview--slicer-org.netlify.app/download.html".